test/spanner/1_cpu_load/requirements.txt
google-cloud-spanner==3.40.0
google-cloud-monitoring==2.18.0
psutil==6.1.0
